+++ src/attr.rs
@@ -1,5 +1,6 @@
 //! [`Attribute`](syn::Attribute) parsing for items, variants and fields.
 
+mod crate_;
 mod default;
 mod field;
 mod incomparable;
@@ -12,6 +13,7 @@ mod zeroize_fqs;
 #[cfg(feature = "zeroize")]
 pub use self::zeroize_fqs::ZeroizeFqs;
 pub use self::{
+	crate_::parse_crate,
 	default::Default,
 	field::FieldAttr,
 	incomparable::Incomparable,