+++ src/error.rs
@@ -167,7 +167,8 @@ impl Error {
 		)
 	}
 
-	/// Invalid value for the `derive_where` or `Zeroize` `crate` option.
+	/// Invalid value for the `derive_where`, `serde` or `Zeroize` `crate`
+	/// option.
 	pub fn path(span: Span, parse_error: syn::Error) -> syn::Error {
 		syn::Error::new(span, format!("expected path, {}", parse_error))
 	}
@@ -189,6 +190,24 @@ impl Error {
 				Self::trait_list()
 			),
 		)
+	}
+
+	/// Requires crate feature `serde`.
+	#[cfg(not(feature = "serde"))]
+	pub fn serde_feature(span: Span) -> syn::Error {
+		syn::Error::new(span, "requires crate feature `serde`")
+	}
+
+	/// Requires crate feature `zeroize`.
+	#[cfg(not(feature = "zeroize"))]
+	pub fn zeroize_feature(span: Span) -> syn::Error {
+		syn::Error::new(span, "requires crate feature `zeroize`")
+	}
+
+	/// Requires crate feature `zeroize-on-drop`.
+	#[cfg(all(feature = "zeroize", not(feature = "zeroize-on-drop")))]
+	pub fn zeroize_on_drop_feature(span: Span) -> syn::Error {
+		syn::Error::new(span, "requires crate feature `zeroize-on-drop`")
 	}
 
 	/// Invalid delimiter in `derive_where` attribute for
@@ -282,6 +301,18 @@ impl Error {
 		syn::Error::new(skip_clone, "Cannot skip `Clone` while deriving `Copy`")
 	}
 
+	/// Unsupported `serde(...)` without deriving `De/Serialize`.
+	#[cfg(feature = "serde")]
+	pub fn serde_without_serde(serde: Span) -> syn::Error {
+		syn::Error::new(serde, "Found unused `#[serde(...)]`")
+	}
+
+	/// Conflicting `serde(bound ...)` when deriving `De/Serialize`.
+	#[cfg(feature = "serde")]
+	pub fn serde_bound(serde: Span) -> syn::Error {
+		syn::Error::new(serde, "Found conflicting `#[serde(bound ...)]`")
+	}
+
 	/// List of available [`Trait`](crate::Trait)s.
 	fn trait_list() -> String {
 		[
@@ -289,14 +320,14 @@ impl Error {
 			"Copy",
 			"Debug",
 			"Default",
+			"Deserialize",
 			"Eq",
 			"Hash",
 			"Ord",
 			"PartialEq",
 			"PartialOrd",
-			#[cfg(feature = "zeroize")]
+			"Serialize",
 			"Zeroize",
-			#[cfg(feature = "zeroize")]
 			"ZeroizeOnDrop",
 		]
 		.join(", ")
@@ -304,15 +335,7 @@ impl Error {
 
 	/// List of available [`SkipGroup`](crate::SkipGroup)s.
 	fn skip_group_list() -> String {
-		[
-			"Clone",
-			"Debug",
-			"EqHashOrd",
-			"Hash",
-			#[cfg(feature = "zeroize")]
-			"Zeroize",
-		]
-		.join(", ")
+		["Clone", "Debug", "EqHashOrd", "Hash", "Zeroize"].join(", ")
 	}
 
 	/// Unsupported `Zeroize` option if [`Zeroize`](https://docs.rs/zeroize/latest/zeroize/trait.Zeroize.html) isn't implemented.