+++ src/trait_/zeroize_on_drop.rs
@@ -5,8 +5,8 @@ use std::{borrow::Cow, iter, ops::Deref};
 use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use syn::{
-	punctuated::Punctuated, spanned::Spanned, Expr, ExprLit, ExprPath, Ident, ImplGenerics, Lit,
-	Meta, Path, Result, Token, TypeGenerics, WhereClause,
+	punctuated::Punctuated, spanned::Spanned, Attribute, DeriveInput, Expr, ExprLit, ExprPath,
+	Ident, ImplGenerics, Lit, Meta, Path, Result, Token, TypeGenerics, WhereClause,
 };
 
 use crate::{util, DeriveTrait, DeriveWhere, Error, Item, SplitGenerics, Trait, TraitImpl};
@@ -34,9 +34,16 @@ impl TraitImpl for ZeroizeOnDrop {
 		})
 	}
 
-	fn parse_derive_trait(_span: Span, list: Punctuated<Meta, Token![,]>) -> Result<DeriveTrait> {
-		// This is already checked in `DeriveTrait::from_stream`.
-		debug_assert!(!list.is_empty());
+	fn parse_derive_trait(
+		_: &[Attribute],
+		_: Span,
+		list: Option<Punctuated<Meta, Token![,]>>,
+	) -> Result<DeriveTrait> {
+		let list = if let Some(list) = list {
+			list
+		} else {
+			return Ok(Self::default_derive_trait());
+		};
 
 		let mut crate_ = None;
 		#[cfg_attr(not(feature = "zeroize-on-drop"), allow(unused_mut))]
@@ -45,16 +52,18 @@ impl TraitImpl for ZeroizeOnDrop {
 		for meta in list {
 			match &meta {
 				Meta::Path(path) => {
-					#[cfg(feature = "zeroize-on-drop")]
 					if path.is_ident("no_drop") {
 						// Check for duplicate `no_drop` option.
+						#[cfg(feature = "zeroize-on-drop")]
 						if !no_drop {
 							no_drop = true;
+							continue;
 						} else {
 							return Err(Error::option_duplicate(path.span(), "no_drop"));
 						}
 
-						continue;
+						#[cfg(not(feature = "zeroize-on-drop"))]
+						return Err(Error::zeroize_on_drop_feature(path.span()));
 					}
 
 					return Err(Error::option_trait(path.span(), Self::as_str()));
@@ -109,6 +118,8 @@ impl TraitImpl for ZeroizeOnDrop {
 
 	fn impl_item(
 		&self,
+		_: Option<&Path>,
+		_: &DeriveInput,
 		imp: &ImplGenerics<'_>,
 		ident: &Ident,
 		ty: &TypeGenerics<'_>,