+++ test-crates/ensure-no-std/src/lib.rs
@@ -8,5 +8,7 @@ use core::marker::PhantomData;
 use derive_where::derive_where;
 
 #[derive_where(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
+#[cfg_attr(feature = "serde", derive_where(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde", serde(crate = "serde_"))]
 #[cfg_attr(feature = "zeroize", derive_where(Zeroize))]
 pub struct Test<T>(PhantomData<T>);