+++ test-crates/minimal-versions/src/lib.rs
@@ -6,5 +6,6 @@ use core::marker::PhantomData;
 use derive_where::derive_where;
 
 #[derive_where(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
+#[cfg_attr(feature = "serde", derive_where(Deserialize, Serialize))]
 #[cfg_attr(feature = "zeroize", derive_where(Zeroize))]
 pub struct Test<T>(PhantomData<T>);