+++ tests/ui/zeroize/item_option_syntax.rs
@@ -1,8 +1,0 @@
-use std::marker::PhantomData;
-
-use derive_where::derive_where;
-
-#[derive_where = "invalid"]
-struct WrongAttributeSyntax<T>(PhantomData<T>);
-
-fn main() {}